gen/traffic-generator/udaan/requirements.txt
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
idna==2.9
loguru==0.4.1
numpy==1.18.1
pycodestyle==2.5.0
python-engineio==3.11.2
python-socketio==4.4.0
requests==2.23.0
six==1.14.0
urllib3==1.25.8
websocket-client==0.57.0
